move components to files
